upgrade CFileItem startOffset and endOffset from int to int64_t

+++ xbmc/Application.cpp
@@ -2248,7 +2248,7 @@ bool CApplication::OnAction(const CAction &action)
       std::string player = CPlayerCoreFactory::GetInstance().SelectPlayerDialog(players);
       if (!player.empty())
       {
-        item.m_lStartOffset = (int)(GetTime() * 75);
+        item.m_lStartOffset = static_cast<int64_t>(GetTime() * 75);
         PlayFile(std::move(item), player, true);
       }
     }
@@ -3010,7 +3010,7 @@ PlayBackRet CApplication::PlayStack(const CFileItem& item, bool bRestart)
   if (!m_pStackHelper->InitializeStack(item))
     return PLAYBACK_FAIL;
 
-  int startoffset = m_pStackHelper->InitializeStackStartPartAndOffset(item);
+  int64_t startoffset = m_pStackHelper->InitializeStackStartPartAndOffset(item);
 
   m_itemCurrentFile.reset(new CFileItem(item));
   CFileItem selectedStackPart = m_pStackHelper->GetCurrentStackPartFileItem();
@@ -4407,7 +4407,7 @@ void CApplication::Restart(bool bSamePosition)
   std::string state = m_pPlayer->GetPlayerState();
 
   // set the requested starttime
-  m_itemCurrentFile->m_lStartOffset = (long)(time * 75.0);
+  m_itemCurrentFile->m_lStartOffset = static_cast<int64_t>(time * 75.0);
 
   // reopen the file
   if ( PlayFile(*m_itemCurrentFile, "", true) == PLAYBACK_OK )
@@ -4625,7 +4625,7 @@ void CApplication::SeekTime( double dTime )
       { // seeking to a new file
         m_pStackHelper->SetStackPartCurrentFileItem(partNumberToPlay);
         CFileItem *item = new CFileItem(m_pStackHelper->GetCurrentStackPartFileItem());
-        item->m_lStartOffset = (static_cast<uint64_t>(dTime * 1000.0) - startOfNewFile) * 75 / 1000;
+        item->m_lStartOffset = (static_cast<int64_t>(dTime * 1000.0) - startOfNewFile) * 75 / 1000;
         // don't just call "PlayFile" here, as we are quite likely called from the
         // player thread, so we won't be able to delete ourselves.
         CApplicationMessenger::GetInstance().PostMsg(TMSG_MEDIA_PLAY, 1, 0, static_cast<void*>(item));
